+++ src/merge/generic_packetizer.h
@@ -59,6 +59,7 @@ protected:
   // by set_headers().
   int m_hserialno, m_htrack_type, m_htrack_min_cache, m_htrack_max_cache;
   int64_t m_htrack_default_duration;
+  bool m_htrack_default_duration_indicates_fields;
   bool m_default_duration_forced;
   bool m_default_track_warning_printed;
   uint64_t m_huid;
@@ -187,7 +188,7 @@ public:
 
   virtual void set_track_min_cache(int min_cache);
   virtual void set_track_max_cache(int max_cache);
-  virtual void set_track_default_duration(int64_t default_duration);
+  virtual void set_track_default_duration(int64_t default_duration, bool force = false);
   virtual void set_track_max_additionals(int max_add_block_ids);
   virtual int64_t get_track_default_duration() const;
   virtual void set_track_forced_flag(bool forced_track);